Add more field mapping

+++ src/main/java/com/gd/termcounter/job/Job.java
@@ -22,8 +22,8 @@ public class Job {
     private String location;
     @Column(columnDefinition = "TEXT")
     private String url;
-    private Double salaryMin;
-    private Double salaryMax;
+    private int salaryMin;
+    private int salaryMax;
 
     public String getKey() {
         return key;
@@ -73,19 +73,19 @@ public class Job {
         this.location = location;
     }
 
-    public Double getSalaryMin() {
+    public int getSalaryMin() {
         return salaryMin;
     }
 
-    public void setSalaryMin(Double salaryMin) {
+    public void setSalaryMin(int salaryMin) {
         this.salaryMin = salaryMin;
     }
 
-    public Double getSalaryMax() {
+    public int getSalaryMax() {
         return salaryMax;
     }
 
-    public void setSalaryMax(Double salaryMax) {
+    public void setSalaryMax(int salaryMax) {
         this.salaryMax = salaryMax;
     }
 }
